requirements.txt
appdirs==1.4.3
awscli==1.16.235
boto==2.49.0
boto3==1.10.28
botocore==1.13.28
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.3.9
ConfigArgParse==1.2.3
cycler==0.10.0
distlib==0.3.0
docutils==0.15.2
filelock==3.0.12
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.2
geventhttpclient==1.4.3
google-api-core==1.22.1
google-auth==1.20.1
google-cloud-speech==1.3.2
googleapis-common-protos==1.52.0
greenlet==0.4.16
grpcio==1.31.0
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.9.4
kiwisolver==1.2.0
locust==1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
msgpack==1.0.0
Naked==0.1.31
numpy==1.18.2
pandas==1.0.3
protobuf==3.13.0
psutil==5.7.0
pyasn1==0.4.7
pyasn1-modules==0.2.8
pycairo==1.18.1
pycrypto==2.6.1
PyGObject==3.34.0
pyparsing==2.4.7
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==19.0.1
requests==2.23.0
rsa==3.4.2
s3transfer==0.2.1
shellescape==3.8.1
six==1.12.0
urllib3==1.25.3
virtualenv==20.0.16
Werkzeug==1.0.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
